Match the live matchup page's chip depth to the archived one
An archived page shows whatever rank the record book holds — the derived lists
are 25 long and #20 has really been rendered — while the live page capped at
five, because it borrowed the HOME STRIP's thresholds. So a #12 blowout got a
chip once archived and nothing at all on the night, which is the same
inconsistency as the missing banner, one layer down.

`getRecordThresholds(depth?)` now takes the depth from the caller. The two
matchup pages pass none and rank against the full list; the home strip and the
season page pass `MARK_DEPTH` and keep their top five, which is right for a
card — chips are worth reading there precisely because most weeks none has one.

Wording followed for the same reason. `RecordMark.long` carries the archived
page's phrasing — "#7 closest win", not the strip's "#7 closest" — so one game
cannot read two ways depending on whether the week has been archived.

Tonight's real games are the check. Den Ops 149.16-148.60, a 0.56 margin:
nothing on the strip, "#7 closest win" on the page. Masterbatters
182.50-121.10: nothing on the strip, "#18 highest score" on the page.

ON THE BASELINE, which prompted this: no change was needed. `derive` builds
records from every finalized week, including those of the season being played,
so by week 10 weeks 1-9 of that season are already in the cut lines along with
every previous year. Verified — den-ops' record lists draw on 2019-2025 today
and 2026 joins the moment its first week archives. The only gap is the week
being played, which cannot rank against itself.

All 1140 archived pages render identical visible text.

+++ components/matchup-preview.tsx
@@ -237,7 +237,8 @@ export function MatchupPreview({
         items={marks.map((m) => {
           const who = m.side === "a" ? a : m.side === "b" ? b : null;
           return {
-            short: m.short,
+            // The PAGE wording, so an archived copy of this game reads the same.
+            short: m.long,
             full: m.full,
             tone: m.tone,
             titleSuffix: who ? ` — ${name(who)}` : "",
